# public/apiDocsV1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "fedora-backend",
    "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
    "contact": {
      "email": ""
    },
    "version": "1.0.0",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api/v1",
...
  "schemes": ["http", "https"],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded",
    "multipart/form-data"
  ],
  "produces": ["application/json"],
  "paths": {
...
    "/grievances/create": {
      "post": {
        "deprecated": false,
        "tags": ["Grievances"],
        "description": "Grievances ",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
             "description": "access token",
            "required": true,
             "type": "string"
           },
           {
            "name": "description",
            "in": "formData",
            "description": "description",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": ["Username: ", "Password: "]
          }
        ]
      }
    },
    "/grievances/list": {
      "get": {
        "deprecated": false,
        "tags": ["Grievances"],
        "description": "Grievances",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
             "description": "access token of admin ",
            "required": true,
             "type": "string"
           },
           {
            "name": "search",
            "in": "query",
            "description": "search by name and mobile number ",    
            "type": "string"
          },
          {
            "name": "date",
            "in": "query",
            "description": "date",    
            "type": "string"
          },
          {
            "name" : "status",
            "in": "query",
            "description": "status 1-pending ,2 - watched , 3 - actionTaken ",         
            "enum": [
              1,
              2,
              3
            ],
            "type": "integer"
          },
          {
            "name": "count",
            "in": "query",
            "description": " count",          
            "type": "number"
          },
          {
            "name": "pageNo",
            "in": "query",
            "description": " pageNo",          
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": ["Username: ", "Password: "]
          }
        ]
      }
    },
    "/grievances/edit": {
      "put": {
        "deprecated": false,
        "tags": ["Grievances"],
        "description": "Customer Account",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "access token of admin ",
            "required": true,
            "type": "string"
          },
          {
            "name": "grievanceId",
            "in": "formData",
            "description": "serviceId",
            "required": true,
            "type": "string"
          },
          {
            "name": "status",
            "in": "formData",
            "description": " serviceStatus",          
            "type": "number"
          }
         
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": ["Username: ", "Password: "]
          }
        ]
      }
    },
    "/grievances/delete": {
      "delete": {
        "deprecated": false,
        "tags": ["Grievances"],
        "description": "deleteService",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
             "description": "access token",
            "required": true,
             "type": "string"
          },
          {
            "name": "grievanceId",
            "in": "query",
            "description": " serviceId",
            "required": false,            
            "type": "string"
          },
          {
            "name": "isDeleted",
            "in": "query",
            "description": "1-soft delete",
            "enum": [
              1
            ],
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": ["Username: ", "Password: "]
          }
        ]
      }
    },
...
  }
...
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    }
  }
}